pyproject.toml
[tool.poetry]
name = "chatbot-lab"
version = "0.1.0"
description = ""
authors = ["Daniel Vergara <dvergara@ghgsat.com>"]
readme = "README.md"
packages = [{include = "chatbot_lab"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
openai = "^0.27.2"
transformers = "^4.27.2"
faiss-cpu = "^1.7.3"
langchain = "^0.0.120"
python-dotenv = "^1.0.0"
click = "^8.1.3"
bs4 = "^0.0.1"
unstructured = "^0.5.6"
pypdf = "^3.6.0"
jsonlines = "^3.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
